rework of ocl handling and choco logic

+++ src/main/java/org/cloudiator/ocl/HardwareBasedPriceFunction.java
@@ -10,8 +10,8 @@ public class HardwareBasedPriceFunction implements PriceFunction {
   @Override
   public double calculatePricing(Cloud cloud, Hardware hardware, Location location, Image image) {
     double price = 0;
-    price += hardware.getCores().doubleValue();
-    price += hardware.getRam().doubleValue() / 1000;
+    price += hardware.getCores();
+    price += hardware.getRam() / 1000;
 
     return price;
   }
